Rename ballerina update tool to ballerina command
