+++ ballerina-command/src/main/java/org/ballerinalang/command/BLauncherCommand.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update;
+package org.ballerinalang.command;
 
 import picocli.CommandLine;
