+++ ballerina-command/src/main/java/org/ballerinalang/command/BallerinaCliCommands.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update;
+package org.ballerinalang.command;
 
 /**
  * Ballerina CLI commands.
@@ -24,7 +24,7 @@ public class BallerinaCliCommands {
 
     public static final String DEFAULT = "default";
     public static final String HELP = "help";
-    public static final String UPDATE = "update";
+    public static final String UPDATE = "command";
     public static final String DIST = "dist";
     public static final String FETCH = "fetch";
     public static final String LIST = "list";
