+++ ballerina-command/src/main/java/org/ballerinalang/command/Main.java
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update;
+package org.ballerinalang.command;
 
-import org.ballerinalang.update.cmd.DefaultCommand;
-import org.ballerinalang.update.cmd.DistributionCommand;
-import org.ballerinalang.update.cmd.FetchCommand;
-import org.ballerinalang.update.cmd.HelpCommand;
-import org.ballerinalang.update.cmd.ListCommand;
-import org.ballerinalang.update.cmd.PullCommand;
-import org.ballerinalang.update.cmd.RemoveCommand;
-import org.ballerinalang.update.cmd.UpdateCommand;
-import org.ballerinalang.update.cmd.UseCommand;
+import org.ballerinalang.command.cmd.DefaultCommand;
+import org.ballerinalang.command.cmd.DistributionCommand;
+import org.ballerinalang.command.cmd.FetchCommand;
+import org.ballerinalang.command.cmd.HelpCommand;
+import org.ballerinalang.command.cmd.ListCommand;
+import org.ballerinalang.command.cmd.PullCommand;
+import org.ballerinalang.command.cmd.RemoveCommand;
+import org.ballerinalang.command.cmd.UpdateCommand;
+import org.ballerinalang.command.cmd.UseCommand;
 import picocli.CommandLine;
 
 import java.io.PrintStream;
