+++ ballerina-command/src/main/java/org/ballerinalang/command/cmd/HelpCommand.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update.cmd;
+package org.ballerinalang.command.cmd;
 
-import org.ballerinalang.update.BLauncherCommand;
-import org.ballerinalang.update.BallerinaCliCommands;
+import org.ballerinalang.command.BLauncherCommand;
+import org.ballerinalang.command.BallerinaCliCommands;
 
 import picocli.CommandLine;
