+++ ballerina-command/src/main/java/org/ballerinalang/command/cmd/PullCommand.java
@@ -14,51 +14,52 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update.cmd;
+package org.ballerinalang.command.cmd;
 
-import org.ballerinalang.update.BLauncherCommand;
-import org.ballerinalang.update.BallerinaCliCommands;
-import org.ballerinalang.update.util.ToolUtil;
+import org.ballerinalang.command.BLauncherCommand;
+import org.ballerinalang.command.BallerinaCliCommands;
+import org.ballerinalang.command.util.ToolUtil;
 import picocli.CommandLine;
 
 import java.io.PrintStream;
 import java.util.List;
 
 /**
- * This class represents the "Fetch" command and it holds arguments and flags specified by the user.
+ * This class represents the "Pull" command and it holds arguments and flags specified by the user.
  */
-@CommandLine.Command(name = "fetch", description = "Fetch Ballerina distribution")
-public class FetchCommand extends Command implements BLauncherCommand {
+@CommandLine.Command(name = "pull", description = "Pull Ballerina distribution")
+public class PullCommand extends Command implements BLauncherCommand {
 
     @CommandLine.Parameters(description = "Command name")
-    private List<String> fetchCommands;
+    private List<String> pullCommands;
 
     @CommandLine.Option(names = {"--help", "-h", "?"}, hidden = true)
     private boolean helpFlag;
 
     private CommandLine parentCmdParser;
 
-    public FetchCommand(PrintStream printStream) {
+    public PullCommand(PrintStream printStream) {
         super(printStream);
     }
 
 
     public void execute() {
         if (helpFlag) {
-            printUsageInfo("dist-" + BallerinaCliCommands.FETCH);
+            printUsageInfo("dist-" + BallerinaCliCommands.PULL);
             return;
         }
 
-        if (fetchCommands == null) {
+        if (pullCommands == null) {
             //    throw LauncherUtils.createUsageExceptionWithHelp("distribution is not provided");
-        } else if (fetchCommands.size() == 1) {
-            ToolUtil.downloadDistribution(getPrintStream(), fetchCommands.get(0), true);
+        } else if (pullCommands.size() == 1) {
+            ToolUtil.downloadDistribution(getPrintStream(), pullCommands.get(0), false);
+            ToolUtil.use(getPrintStream(), pullCommands.get(0));
             return;
-        } else if (fetchCommands.size() > 1) {
+        } else if (pullCommands.size() > 1) {
             //    throw LauncherUtils.createUsageExceptionWithHelp("too many arguments given");
         }
 
-        String userCommand = fetchCommands.get(0);
+        String userCommand = pullCommands.get(0);
         if (parentCmdParser.getSubcommands().get(userCommand) == null) {
             //    throw LauncherUtils.createUsageExceptionWithHelp("unknown command " + userCommand);
         }
@@ -76,7 +77,7 @@ public class FetchCommand extends Command implements BLauncherCommand {
 
     @Override
     public void printUsage(StringBuilder out) {
-        out.append("  ballerina dist fetch\n");
+        out.append("  ballerina dist pull\n");
     }
 
     @Override
