+++ ballerina-command/src/main/java/org/ballerinalang/command/util/Version.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ballerinalang.update.util;
+package org.ballerinalang.command.util;
 
 /**
  * Ballerina version tools.
