+++ ballerina-command/src/main/resources/META-INF/tool.properties
