+++ ballerina-command/src/main/resources/cli-help/ballerina-list.help
