+++ ballerina-command/src/main/resources/cli-help/ballerina-remove.help
